Implement game loop with console clearing and enhanced attack feedback
