+++ app/classes/Character.py
@@ -73,7 +73,10 @@ class Character:
             damage = 20
             enemy.hp -= damage
             if enemy.hp <= 0:
+                print(f"{self.name} attaque {enemy.name} et inflige {damage} HP de dégâts. \033[91m{enemy.name} est mort.\033[0m🪦")
                 enemy.hp = 0
+                return enemy.name
+
             print(f"{self.name} attaque {enemy.name} et inflige {damage} HP de dégâts. {enemy.name} a maintenant {enemy.hp} HP.")
             return enemy.name
         
@@ -82,7 +85,9 @@ class Character:
             damage = random.randint(1, 10)
             self.hp -= damage
             if enemy.hp <= 0:
+                print(f"{self.name} attaque {enemy.name} mais l'attaque est reflétée et inflige {damage} HP de dégâts à {self.name}.  \033[91m{enemy.name} est mort.\033[0m🪦")
                 enemy.hp = 0
+                return None
             print(f"{self.name} attaque {enemy.name} mais l'attaque est reflétée et inflige {damage} HP de dégâts à {self.name}. {self.name} a maintenant {self.hp} HP.")
             return None
         
@@ -90,7 +95,9 @@ class Character:
         damage = random.randint(1, 10)
         enemy.hp -= damage
         if enemy.hp <= 0:
+            print(f"{self.name} attaque {enemy.name} et inflige {damage} HP de dégâts. \033[91m{enemy.name} est mort.\033[0m🪦")
             enemy.hp = 0
+            return enemy.name
         print(f"{self.name} attaque {enemy.name} et inflige {damage} HP de dégâts. {enemy.name} a maintenant {enemy.hp} HP.")
 
         return enemy.name
